+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The main work (all changes without a GitHub username in brackets in the below li
 
 -   @matter/node
     - Enhancement: Matter protocol initialization now runs independently of and after behavior initialization, giving behaviors more flexibility in participating in protocol setup
+    - Enhancement: Each new PASE session now automatically arms the failsafe timer for 60s as required by specs
     - Fix: Fixes withBehaviors() method on endpoints
 
 -   @matter/nodejs-ble
@@ -30,7 +31,11 @@ The main work (all changes without a GitHub username in brackets in the below li
     - FIx: Restores the possibility to cancel a (continuous) discovery for commissionable devices
 
 -   @project-chip/matter.js
-    - Enhancement: Improves Reconnection Handling for devices that use Persisted Subscriptions
+    - Feature: (Breaking) Added Fabric Label for Controller as required property to initialize the Controller
+        including setting the Fabric Label when commissioning and validating and updating the Fabric Label on
+        connection 
+    - Feature: Allows to update the Fabric Label during controller runtime using `updateFabricLabel()` on CommissioningController
+    - Enhancement: Improves Reconnection Handling for devices that use persisted subscriptions
     - Enhancement: Use data type definitions from Model for Controller Device type definitions
 
 ## 0.11.9 (2024-12-11)